finish childStore beta
